Refactor AlbumPlayerView to use AlbumVinylPlayer
Moves the vinyl player component into its own dedicated component. This
simplifies the `AlbumPlayerView` and makes the vinyl player more
reusable.
