+++ source/features/playground/experiments/music-details/AlbumVinylPlayer.module.css
